refactor & feat : PostServlet에 get post전부 만듦
